+++ tsup.config.ts
@@ -1,7 +1,10 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: {
+    index: 'src/index.ts',
+    'cli/index': 'src/cli/index.ts',
+  },
   format: ['esm'],
   dts: true,
   splitting: false,
@@ -17,5 +20,6 @@ export default defineConfig({
     'typebox',
     '@sinclair/typebox',
     'proper-lockfile',
+    'better-sqlite3',
   ],
 });